# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "File Management System API",
    "description": "Backend for a complete file management system: JWT auth with email-verification OTPs, file uploads with content extraction (PDF/DOCX/XLSX/CSV/text/images) and Cloudinary storage, per-user dashboards, admin statistics, and admin user management.\n\nAll authenticated endpoints expect `Authorization: Bearer <accessToken>`. The token payload carries `{ userId, role }` only.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" },
    { "url": "https://your-project.vercel.app", "description": "Vercel deployment (replace with your domain)" }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List users",
        "description": "Admin only. Search matches name OR email (case-insensitive); filter by role, verification status, creation-date range; sort whitelist: createdAt, updatedAt, name, email. Password is stripped from every payload.",
        "operationId": "listUsers",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 }, "example": 1 },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 }, "example": 10 },
          { "name": "search", "in": "query", "schema": { "type": "string", "minLength": 1, "maxLength": 255 }, "description": "Matches name or email (case-insensitive)", "example": "alicia" },
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["USER", "ADMIN"] }, "example": "ADMIN" },
          { "name": "isVerified", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] }, "example": "true" },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" }, "description": "createdAt >= from", "example": "2026-08-01T00:00:00.000Z" },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" }, "description": "createdAt <= to", "example": "2026-08-31T23:59:59.000Z" },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "enum": ["createdAt", "updatedAt", "name", "email"], "default": "createdAt" }, "example": "name" },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" }, "example": "asc" }
        ],
        "responses": {
          "200": {
            "description": "Paginated users",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsersListResult" },
                "example": {
                  "data": [
                    {
                      "id": "7c2e9a41-53fd-4b18-92c6-1de4f0a83bb2",
                      "name": "Alicia B",
                      "email": "alicia-b@suite.dev",
                      "role": "USER",
                      "isVerified": true,
                      "createdAt": "2026-08-25T00:16:54.100Z",
                      "updatedAt": "2026-08-25T00:16:54.100Z"
                    }
                  ],
                  "pagination": { "page": 1, "limit": 10, "total": 1, "totalPages": 1 }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Invalid request payload",
                    "details": [{ "path": "isVerified", "message": "Invalid enum value. Expected 'true' | 'false'" }]
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/users/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "example": "7c2e9a41-53fd-4b18-92c6-1de4f0a83bb2" }
      ],
      "patch": {
        "tags": ["Users"],
        "summary": "Update a user",
        "description": "Admin only. At least one of `name`, `role`, `isVerified` must be provided. Guards: admins cannot change their own role; demoting the last remaining ADMIN returns 409.",
        "operationId": "updateUser",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserInput" },
              "example": { "role": "ADMIN" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserMutationResponse" },
                "example": {
                  "message": "User updated successfully",
                  "user": {
                    "id": "7c2e9a41-53fd-4b18-92c6-1de4f0a83bb2",
                    "name": "Alicia B",
                    "email": "alicia-b@suite.dev",
                    "role": "ADMIN",
                    "isVerified": true,
                    "createdAt": "2026-08-25T00:16:54.100Z",
                    "updatedAt": "2026-08-25T00:18:12.480Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (empty body, bad values)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Invalid request payload",
                    "details": [{ "path": "", "message": "At least one field (name, role, isVerified) is required" }]
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "Changing your own role is forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": { "error": { "code": "FORBIDDEN", "message": "You cannot change your own role" } }
              }
            }
          },
          "404": {
            "description": "Unknown user id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": { "error": { "code": "NOT_FOUND", "message": "User not found" } }
              }
            }
          },
          "409": {
            "description": "Demotion would leave zero admins",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": { "error": { "code": "CONFLICT", "message": "Cannot demote the last remaining admin" } }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete a user",
        "description": "Admin only. Cascades the user's files and verification codes; owned Cloudinary assets are destroyed best-effort. Guards: cannot delete yourself; cannot delete the last remaining ADMIN.",
        "operationId": "deleteUser",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageOnly" },
                "example": { "message": "User deleted successfully" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "Self-deletion forbidden (or non-admin caller)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": { "error": { "code": "FORBIDDEN", "message": "You cannot delete your own account" } }
              }
            }
          },
          "404": {
            "description": "Unknown user id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": { "error": { "code": "NOT_FOUND", "message": "User not found" } }
              }
            }
          },
          "409": {
            "description": "Deletion would leave zero admins",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": { "error": { "code": "CONFLICT", "message": "Cannot delete the last remaining admin" } }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token returned by POST /auth/login. Payload: { userId, role }."
      }
    },
    "schemas": {
...
      "Role": { "type": "string", "enum": ["USER", "ADMIN"] },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "email": { "type": "string", "format": "email" },
          "role": { "$ref": "#/components/schemas/Role" },
          "isVerified": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "name", "email", "role", "isVerified", "createdAt", "updatedAt"]
      },
...
      "UpdateUserInput": {
        "type": "object",
        "minProperties": 1,
        "additionalProperties": false,
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "role": { "$ref": "#/components/schemas/Role" },
          "isVerified": { "type": "boolean" }
        }
      },
      "MessageOnly": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "required": ["message"]
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "minimum": 1 },
          "limit": { "type": "integer", "minimum": 1, "maximum": 100 },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer", "minimum": 1 }
        },
        "required": ["page", "limit", "total", "totalPages"]
      },
...
      "UsersListResult": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
          "pagination": { "$ref": "#/components/schemas/PaginationMeta" }
        },
        "required": ["data", "pagination"]
      },
      "UserMutationResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        },
        "required": ["message", "user"]
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "path": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["path", "message"]
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": ["VALIDATION_ERROR", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "CONFLICT", "FILE_TOO_LARGE", "MULTIPART_ERROR", "UPLOAD_FAILED", "EMAIL_SEND_FAILED", "INTERNAL_SERVER_ERROR"]
              },
              "message": { "type": "string" },
              "details": {
                "type": "array",
                "nullable": true,
                "items": { "$ref": "#/components/schemas/ErrorDetail" },
                "description": "Present only for zod validation failures."
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing, invalid, or expired bearer token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": { "error": { "code": "UNAUTHORIZED", "message": "Authentication required" } }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but insufficient role",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": { "error": { "code": "FORBIDDEN", "message": "Insufficient permissions" } }
          }
        }
      },
...
    }
  }
}